As a temporary trial, split up the Main into BTMain and GPSMain
WHY:
Things keep changing in Main on the GPS side that are affecting the BT side. Instead of trying
to make the MainJS an abstraction that works for both, separate out the two Main's and then
figure out what is similar at that point.

This Commit:
First thing was removing the temporary global for ENPermissions. This wasn't working correctly
anyways after Main was changed.

Second thing was to split the Main.js file into two components for the time being and move the isGPS
logic to the Entry.js to decide which main to load.

Third thing, remove the global variable and connect the BTMain to the ExposureNotificationsContext.

Lastly, replace the onPress on the TracingOff screen depending if it is GPS or BT.